Avoid failing on erroneous lines.
Trial-and-error is painful when testing takes time. Since the proccess of
generating the n-grams database is long and ressources-consuming, it’s
better not stop in the middle just because of one or two malformed lines.
This allows the user to get a complete list of erroneous lines, fix them
all at once and run the script again.